Refac: move module and add todos
