+++ src/careamics/compat/utils/__init__.py
@@ -1,0 +1,7 @@
+"""Deprecated."""
+
+from .path_utils import check_path_exists
+
+__all__ = [
+    "check_path_exists",
+]
