+++ src/careamics/utils/__init__.py
@@ -3,7 +3,6 @@
 __all__ = [
     "BaseEnum",
     "autocorrelation",
-    "check_path_exists",
     "cwd",
     "get_careamics_home",
     "get_logger",
@@ -15,5 +14,4 @@ from .autocorrelation import autocorrelation
 from .base_enum import BaseEnum
 from .context import cwd, get_careamics_home
 from .logging import get_logger
-from .path_utils import check_path_exists
 from .ram import get_ram_size
